Renamed the translator module to tracwiki2githubmd.
